stewart-jason-hw6
