--- hw6/_submissions/drean-patrik/fibonacci/src/fibonacci.js
@@ -1,4 +1,0 @@
-export const getFibNumber = steps => {
-  // TODO
-  return undefined
-}
