--- hw6/_submissions/drean-patrik/fibonacci/src/fibonacci.test.js
@@ -1,11 +1,0 @@
-import { getFibNumber } from './fibonacci'
-
-describe('When getting the fibonacci number', () => {
-  it('should return 0 as the first value', () => {
-    const result = getFibNumber(0)
-    const expected = 0
-
-    expect(result).toEqual(expected)
-  })
-  // TODO: Write more tests
-})
