--- hw6/_submissions/drean-patrik/spinning/README.md
@@ -1,18 +1,0 @@
-# Spinning
-
-## Setup
-
-Copy this folder and paste it under `_submissions/lastname-firstname`.
-
-Now in the new directory, run `npm install` in order to install the required packages.
-
-The code is located in the `src` folder.
-
-Run `npm run test` to verify if tests are passing or not. Remember, you have to be in the directory for this command to work.
-
-## Instructions
-
-Write a function `spinWords(sentence)` that takes in a string of one or more words, and returns the same string, but with all five or more letter words reversed. Strings passed in will consist of only letters and spaces. Spaces will be included only when more than one word is present.
-
-- For example:
-  - `"Stopping Spinning My Words"` should return `"Stop gninnipS My sdroW"`
