--- hw6/_submissions/drean-patrik/spinning/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "jest-template",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "jest"
-  },
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "@babel/core": "^7.6.4",
-    "@babel/preset-env": "^7.6.3",
-    "babel-jest": "^24.9.0",
-    "jest": "^24.9.0"
-  }
-}
