--- hw6/_submissions/drean-patrik/spinning/src/spinning.js
@@ -1,4 +1,0 @@
-export const spinWords = sentence => {
-  // TODO
-  return undefined
-}
