--- hw6/_submissions/drean-patrik/spinning/src/spinning.test.js
@@ -1,11 +1,0 @@
-import { spinWords } from './spinning'
-
-describe('When spinning words', () => {
-  it('should not spin any words if under 5 characters', () => {
-    const result = spinWords('no spin now')
-    const expected = 'no spin now'
-
-    expect(result).toEqual(expected)
-  })
-  // TODO: Add more tests
-})
